consolidate scheduler settings
